requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
python-multipart
google-generativeai
